Hotfix/fix staging branch (#491)
