+++ .env.validator.example
@@ -1,5 +1,7 @@
 # The network UID. If 1 for main, 61 for testnet.
 NETUID = 61
+DEPLOY_API = false
+API_PORT = 8094
 
 # The network name [test, finney, local].
 SUBTENSOR_NETWORK = "test"
