+++ .gitignore
@@ -180,3 +180,4 @@ wandb
 .vscode
 api_keys.json
 prompting/api/api_keys.json
+AutoAWQ
