+++ prompting/base/validator.py
@@ -9,7 +9,6 @@ import torch
 from loguru import logger
 
 from prompting.base.neuron import BaseNeuron
-from prompting.rewards.reward import WeightedRewardEvent
 from prompting.settings import settings
 from prompting.utils.exceptions import MaxRetryError
 from prompting.utils.logging import init_wandb
@@ -44,12 +43,6 @@ class BaseValidatorNeuron(BaseNeuron):
         self.is_running: bool = False
         self.thread: threading.Thread = None
 
-    def _serve_axon(self):
-        """Serve axon to enable external connections"""
-        validator_uid = settings.METAGRAPH.hotkeys.index(settings.WALLET.hotkey.ss58_address)
-        self.axon.serve(netuid=settings.NETUID, subtensor=settings.SUBTENSOR).start()
-        logger.info(f"Serving validator UID {validator_uid} on {self.axon.ip}:{self.axon.port} to chain")
-
     def run(self):
         """
         Initiates and manages the main loop for the miner on the Bittensor network. The main loop handles graceful shutdown on keyboard interrupts and logs unforeseen errors.
@@ -72,10 +65,7 @@ class BaseValidatorNeuron(BaseNeuron):
         # Check that validator is registered on the network.
         self.sync()
 
-        if not settings.NEURON_AXON_OFF:
-            logger.info(f"Running validator on netuid: {settings.NETUID}")
-        else:
-            logger.info(f"Running validator with netuid: {settings.NETUID}")
+        logger.info(f"Running validator with netuid: {settings.NETUID}")
 
         logger.info(f"Validator starting at block: {self.latest_block}")
 
@@ -86,8 +76,9 @@ class BaseValidatorNeuron(BaseNeuron):
 
                 forward_timeout = settings.NEURON_FORWARD_MAX_TIME
                 try:
-                    task = self.loop.create_task(self.forward())
-                    self.loop.run_until_complete(asyncio.wait_for(task, timeout=forward_timeout))
+                    loop = asyncio.get_event_loop()
+                    result = loop.run_until_complete(self.forward())
+                    logger.debug(f"Result of forward loop {result}")
                 except torch.cuda.OutOfMemoryError as e:
                     logger.error(f"Out of memory error: {e}")
                     continue
@@ -113,7 +104,6 @@ class BaseValidatorNeuron(BaseNeuron):
 
         # If someone intentionally stops the validator, it'll safely terminate operations.
         except KeyboardInterrupt:
-            self.axon.stop()
             logger.success("Validator killed by keyboard interrupt.")
             sys.exit()
 
@@ -174,79 +164,5 @@ class BaseValidatorNeuron(BaseNeuron):
     def resync_metagraph(self):
         """Resyncs the metagraph and updates the hotkeys and moving averages based on the new metagraph."""
         logger.info("resync_metagraph()")
-
-        # Copies state of metagraph before syncing.
-        previous_metagraph = copy.deepcopy(settings.METAGRAPH)
-
         # Sync the metagraph.
         settings.METAGRAPH.sync(subtensor=settings.SUBTENSOR)
-
-        # Check if the metagraph axon info has changed.
-        if previous_metagraph.axons == settings.METAGRAPH.axons:
-            return
-
-        logger.info("Metagraph updated, re-syncing hotkeys, dendrite pool and moving averages")
-        logger.info("Metagraph updated, re-syncing hotkeys, dendrite pool and moving averages")
-        # Zero out all hotkeys that have been replaced.
-        for uid, hotkey in enumerate(self.hotkeys):
-            if hotkey != settings.METAGRAPH.hotkeys[uid]:
-                self.scores[uid] = 0  # hotkey has been replaced
-
-        # Check to see if the metagraph has changed size.
-        # If so, we need to add new hotkeys and moving averages.
-        if len(self.hotkeys) < len(settings.METAGRAPH.hotkeys):
-            # Update the size of the moving average scores.
-            new_moving_average = np.zeros((settings.METAGRAPH.n))
-            min_len = min(len(self.hotkeys), len(self.scores))
-            new_moving_average[:min_len] = self.scores[:min_len]
-            self.scores = new_moving_average
-
-        # Update the hotkeys.
-        self.hotkeys = copy.deepcopy(settings.METAGRAPH.hotkeys)
-
-    def update_scores(self, reward_events: list[WeightedRewardEvent]):
-        """Performs exponential moving average on the scores based on the rewards received from the miners."""
-        for r_event in reward_events:
-            # Check if rewards contains NaN values.
-            rewards = r_event.rewards_normalized
-            if any(np.isnan(rewards).flatten()):
-                # if
-                logger.warning(f"NaN values detected in rewards: {rewards}")
-                # Replace any NaN values in rewards with 0.
-                rewards = np.nan_to_num(rewards)
-
-            # Compute forward pass rewards, assumes uids are mutually exclusive.
-            # shape: [ metagraph.n ]
-            step_rewards = np.copy(self.scores)
-            step_rewards[np.array(r_event.uids).astype(int)] = rewards
-            logger.debug(f"Scattered rewards: {rewards}")
-
-            # Update scores with rewards produced by this step.
-            # shape: [ metagraph.n ]
-            alpha = settings.NEURON_MOVING_AVERAGE_ALPHA
-            self.scores = alpha * step_rewards + (1 - alpha) * self.scores
-            self.scores = np.clip(self.scores - settings.NEURON_DECAY_ALPHA, 0, 1)
-            logger.debug(f"Updated moving avg scores: {self.scores}")
-
-    def save_state(self):
-        """Saves the state of the validator to a file."""
-        logger.info("Saving validator state.")
-
-        # Save the state of the validator to file.
-        try:
-            np.savez(settings.SAVE_PATH + "/state.npz", step=self.step, scores=self.scores, hotkeys=self.hotkeys)
-        except Exception as ex:
-            logger.error(f"Couldn't save state to file: {ex}")
-
-    def load_state(self):
-        """Loads the state of the validator from a file."""
-        logger.info("Loading validator state.")
-
-        # Load the state of the validator from file.
-        try:
-            state = np.load(settings.SAVE_PATH + "/state.npz")
-            self.step = state["step"]
-            self.scores = state["scores"]
-            self.hotkeys = state["hotkeys"]
-        except Exception as ex:
-            logger.error(f"Couldn't load state from file: {ex}")
